+++ kinetic/__init__.py
@@ -16,6 +16,10 @@
 
 #@author: Ignacio Corderi
 
+# Logging
+import logging
+logging.basicConfig()
+
 # Protocol version
 from common import local
 
@@ -37,17 +41,24 @@ else:
 #utils
 from utils import buildRange
 
-# client
-from client import Client
-from asyncclient import AsyncClient
+# clients
+from greenclient import Client
+from secureclient import SecureClient
 from threadedclient import ThreadedClient
 
 # common
 from common import KeyRange
 from common import Entry
+from common import Peer
 
 # exceptions
 from common import KineticMessageException
 
-# Admin
-from admin import AdminClient
+# backward compatibility alliases
+AsyncClient = Client
+from kinetic.deprecated.adminclient import AdminClient
+from kinetic import greenclient as client
+# Fake old asyncclient module 
+class AsyncClientCompat(object):
+    AsyncClient = Client   
+asyncclient = AsyncClientCompat()