+++ kinetic/baseasync.py
@@ -16,7 +16,7 @@
 
 #@author: Ignacio Corderi
 
-from client import Client
+import deprecated
 from common import Entry
 import common
 
@@ -27,7 +27,7 @@ import threading
 
 LOG = logging.getLogger(__name__)
 
-class BaseAsync(Client):
+class BaseAsync(deprecated.BlockingClient):
 
     def __init__(self, *args, **kwargs):
         super(BaseAsync, self).__init__(*args, socket_timeout=None, **kwargs)
@@ -132,7 +132,7 @@ class BaseAsync(Client):
     ###
 
 
-    def sendAsync(self, command, value, onSuccess, onError):
+    def sendAsync(self, command, value, onSuccess, onError, no_ack=False):
         if self.faulted: # TODO(Nacho): should we fault through onError on fault or bow up on the callers face?
             self._raise(common.ConnectionFaulted("Can't send message when connection is on a faulted state."), onError)
             return #skip the rest
@@ -152,8 +152,9 @@ class BaseAsync(Client):
         # get sequence
         self.update_header(command)
 
-        # add callback to pending dictionary
-        self._pending[command.header.sequence] = (innerSuccess, onError)
+        if not no_ack:
+            # add callback to pending dictionary
+            self._pending[command.header.sequence] = (innerSuccess, onError)
 
         # transmit
         self.network_send(command, value)
@@ -177,8 +178,14 @@ class BaseAsync(Client):
             except Exception as e2:
                 onError(e2)
 
+        if 'no_ack' in kwargs:
+            send_no_ack = True
+            del kwargs['no_ack']
+        else:
+            send_no_ack = False
+
         header, value = op.build(*args, **kwargs)
-        self.sendAsync(header, value, innerSuccess, innerError)
+        self.sendAsync(header, value, innerSuccess, innerError, send_no_ack)
 
 
     def putAsync(self, onSuccess, onError, *args, **kwargs):