+++ kinetic/common.py
@@ -18,6 +18,7 @@
 
 import kinetic_pb2 as messages
 import eventlet
+import os
 
 MAX_KEY_SIZE = 4*1024
 MAX_VALUE_SIZE = 1024*1024
@@ -26,6 +27,12 @@ DEFAULT_CONNECT_TIMEOUT = 0.1
 DEFAULT_SOCKET_TIMEOUT = 5
 DEFAULT_CHUNK_SIZE = 64*1024
 
+# Env variables
+try: 
+    KINETIC_CONNECT_TIMEOUT = float(os.environ.get('KINETIC_CONNECT_TIMEOUT', DEFAULT_CONNECT_TIMEOUT))
+except:
+    KINETIC_CONNECT_TIMEOUT = DEFAULT_CONNECT_TIMEOUT
+    
 local = messages.Local()
 
 class DeferedValue():
@@ -164,6 +171,14 @@ class ClusterVersionFailureException(KineticMessageException):
         super(ClusterVersionFailureException, self).__init__(status)
         self.cluster_version = cluster_version
 
+class BatchAbortedException(KineticException):
+    def __init__(self, value):
+        super(BatchAbortedException, self).__init__(value)
+        self.failed_operation_index = -1
+
+class BatchCompletedException(KineticClientException):
+    def __init__(self):
+        super(BatchCompletedException, self).__init__('batch completed. no more operations are permitted within this batch.')
 
 class HmacAlgorithms:
     INVALID_HMAC_ALGORITHM = -1 # Must come first, so default is invalid